Fix double-click play and jumpy seeker bugs
Clicking the play button twice in quick succession would pause the
player but not update the pause/play button. This seems to be because
autoplay=on only updates the paused attribute once enough media has
buffered for play to start. Calling play() manually whenever the audio source changes seems to resolve this.

To prevent the range input from jumping about when changing between
tracks, set it back to 0 whenever a new track is played.