Remove startingTrack-based logic
It's necessary to maintain the logic around having a track selected vs
having a track playing, but it doesn't require two variables.

Instead, currentTrack can be set when the tracks are loaded. However,
the <audio> element's src is only set once a track is actually played.
If you need to check whether _any_ tracks have played yet, you can check
whether this src is an empty string.

Part of fixing this involves undoing the statement that automatically
sets the src whenever the currentTrack changes. Since the currentTrack
is always defined now, the src would be set immediately and clients will
load the audio file. This is undesirable - audio should only be loaded
once play is actually triggered. Luckily, it makes sense to bundle the
logic into the main play() function.