Substitute origin into embed client-side
Some details are better obtained client-side, rather than by the worker.
Namely, the origin the worker is being served from. It's unreliable to
fetch this from headers or request data, since the value doesn't match
to the actual origin when using wrangler dev (instead of localhost, it's
the deployed URL).